requirements.txt
asgiref==3.9.1
Django==5.2.6
django-filter==25.1
django-jalali==7.4.0
django-js-asset==3.1.2
django-mptt==0.18.0
djangorestframework==3.16.1
jalali_core==1.0.0
jdatetime==5.2.0
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
